requirements.txt
Django==1.10.5
djangorestframework==3.6.2
django-environ==0.4.1
